fix(plugin-api): keep terminal run filters internal
