+++ test-fixtures/plugin-api-baseline/shared/pluginApiTypes.d.ts
@@ -90,13 +90,6 @@ export interface TerminalCommandRunHandle {
     run: TerminalCommandRun;
     completed: Promise<TerminalCommandRun>;
 }
-export interface TerminalCommandRunFilter {
-    projectId?: string;
-    workspaceId?: string;
-    terminalId?: string;
-    statuses?: TerminalCommandRunStatus[];
-    metadata?: Record<string, string>;
-}
 export type PiWebServiceComponent = "web" | "sessiond";
 export type PiWebStatusSeverity = "info" | "warning" | "error";
 export type PiWebInstallationKind = "pi-package" | "npm-global" | "local" | "docker" | "unknown";
